backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn[standard]
websockets
opencv-python
ultralytics
numpy
paho-mqtt
sqlalchemy
python-dotenv
Pillow
mediapipe
python-dotenv

# ── Persistence layer ───────────────────────────────────
sqlalchemy[asyncio]     
asyncpg                
alembic              
pydantic-settings
aiofiles
python-multipart
 